requirements.txt
Django==2.1
psycopg2-binary==2.7.7
django-appconf==1.0.2
django-cors-headers==2.2.0
django-imagekit==4.0.2
django-model-utils==3.1.2
djangorestframework==3.8.2
django-extra-fields==1.0.0
git+https://github.com/paulorobertocruz/drf-extra-fields
django-widget-tweaks==1.4.2
lazy-object-proxy==1.3.1
pilkit==2.0
Pillow==5.1.0
python-decouple==3.1
dj-database-url==0.5.0
pytz==2018.4
six==1.11.0
wrapt==1.10.11
gunicorn==19.9.0
filemagic==1.6
django-rest-swagger==2.2.0
djangorestframework-jwt==1.11.0
